+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - [IMPROVEMENT] - Fix value normalization for dirty comparison (when comparing arrays)
 - [FEATURE] - Add "isSaveError" state
+- [BUGFIX] - Fix when should model properties sync with model
 
 # 0.2.6
 - [FEATURE] - Add support for readonly properties that can't be in dirty state